refactor: kafka 소비자 위치 변경 및 import 경로 업데이트
