+++ src/v1/kafka/consumer.ts
@@ -1,8 +1,8 @@
 import { Namespace } from 'socket.io';
 import { TOPICS, GROUP_IDS, USER_STATUS_EVENTS, FRIEND_EVENTS } from './constants.js';
 import { kafka } from '../../plugins/kafka.js';
-import FriendConsumer from './friend.consumer.js';
-import UserStatusConsumer from './user-status.consumer.js';
+import FriendConsumer from './consumers/friend.consumer.js';
+import UserStatusConsumer from './consumers/user-status.consumer.js';
 import { userStatus } from '../sockets/status/status.schema.js';
 import { friendAddMessage, friendBlockMessage, userStatusMessage } from './messages.schema.js';
